DogPush doesn't run on python 3
Updated dependency versions, updated print statements and changed
iteration iteration operations on various dicts (in the creation
of option_defaults for example) in order for the module to run
on python 3.7.something.

There have been no attempts to improve much else, since that is not my
goal, though it is evident that much work is required to bring the
module to a state by which it could be said to "work reliably".

To me, getting the module to run is enough; effort is better spent
creating a cdk module or somethin' idk byeeeeeeeeeeeeee